rename: pages 폴더명 page로 변경
